Splitting spec_helper into smaller files
