+++ spec/login_helper.rb
@@ -1,0 +1,20 @@
+Mumukit::Login.configure do |config|
+  config.auth0 = struct
+  config.saml = struct
+
+  config.mucookie_domain = '.localmumuki.io'
+  config.mucookie_secret_key = 'abcde1213456123456'
+  config.mucookie_secret_salt = 'mucookie test secret salt'
+  config.mucookie_sign_salt = 'mucookie test sign salt'
+end
+
+OmniAuth.config.test_mode = true
+OmniAuth.config.mock_auth[:developer] =
+  OmniAuth::AuthHash.new provider: 'developer',
+                         uid: 'johndoe@test.com',
+                         credentials: {},
+                         info: {first_name: 'John', last_name: 'Doe', name: 'John Doe', nickname: 'johndoe'}
+
+def set_current_user!(user)
+  allow_any_instance_of(ApplicationController).to receive(:current_user_uid).and_return(user.uid)
+end
